requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai>=1.60.0
python-dotenv==1.0.1
pydantic==2.9.2
supabase>=2.0.0
google-generativeai>=0.8.0

Faker>=20.0.0
